requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2024.8.30
charset-normalizer==3.4.0
cloudpickle==3.1.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
docstring_parser==0.16
Farama-Notifications==0.0.4
fonttools==4.54.1
gymnasium==1.0.0
idna==3.10
ipython-genutils==0.2.0
joblib==1.4.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
numpy==2.1.2
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
scikit-learn==1.5.2
scipy==1.14.1
shtab==1.7.1
six==1.16.0
threadpoolctl==3.5.0
traitlets_pcse==5.0.0.dev0
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
urllib3==2.2.3
